Revert "Change data type from float to float32 in indexed_dataset.py, teacher_out_dataset.py, resampling_dataset.py, and fed_utils.py"
This reverts commit 8f38a686a018a2592db3c3e7a543fc315f777f35.

+++ examples/code-backdoor/Birnn_Transformer/ncc/data/kd/teacher_out_dataset.py
@@ -19,7 +19,7 @@ class TeacherOutDataset(IndexedCachedDataset):
         super().__init__(prefix, fix_lua_indexing=False)
 
     @staticmethod
-    def save_bin(prefix, data_list, dtype=np.float32):
+    def save_bin(prefix, data_list, dtype=np.float):
         bin_path = prefix + '.mmap'
         idx_path = prefix + '.idx'
         builder = TeacherOutDatasetBuilder(bin_path, dtype)
